feat: 예약이 존재하는 시간을 삭제할 수 없도록 예외 처리

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -34,6 +34,9 @@ public class ReservationTimeService {
     }
 
     public void delete(Long id) {
+        if (reservationTimeDao.existsByTimeId(id)) {
+            throw new IllegalArgumentException("해당 시간에 예약이 존재하여 삭제할 수 없습니다.");
+        }
         reservationTimeDao.delete(id);
     }
 }
